backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic
email-validator
python-multipart
PyJWT
bcrypt
httpx
Pillow
sentence-transformers
alembic
pytest
